Fix GKE Windows CODEOWNERS again
For some reason on
https://github.com/GoogleCloudPlatform/kubernetes-engine-samples/pull/271
GitHub is only detecting mcshooter as a code owner, and not ibabou.
Michelle is on leave and can't approve that PR, so I'm trying to fix the
CODEOWNERS by removing mcshooter and liurupeng. liurupeng is on the team
but is not a GoogleCloudPlatform org member yet, so perhaps that's
tripping up the GitHub owner detector somehow. I don't see any other
good reason why ibabou would not have owner approval power with the
current file.