Add sample application exporting Prometheus metrics.
This sample includes a simple application that generates Prometheus
metrics and exports them to Google Cloud Monitoring using Google
Cloud Managed Service for Prometheus.

It superseeds the 'prometheus-to-sd' sample which uses the deprecated
'prometheus-to-sd' sidecar container. The simple application generating
Prometheus metrics (and its build configuration) has been moved to this
sample.